# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SpaceLink API Documents",
        "description": "Hệ thống API backend viết bằng Laravel 12",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/client/checkout": {
            "post": {
                "tags": [
                    "Client - Checkout"
                ],
                "summary": "Đặt hàng từ giỏ hàng hiện tại",
                "operationId": "f61fbaf721caf05ca415124401f56c94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shipping_address_id",
                                    "payment_method"
                                ],
                                "properties": {
                                    "shipping_address_id": {
                                        "description": "ID địa chỉ giao hàng của user",
                                        "type": "integer"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cod",
                                            "vnpay",
                                            "momo",
                                            "bank_transfer"
                                        ]
                                    },
                                    "voucher_code": {
                                        "description": "Mã voucher (không bắt buộc)",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "note": {
                                        "description": "Ghi chú đơn hàng",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đặt hàng thành công"
                    },
                    "400": {
                        "description": "Giỏ hàng rỗng hoặc lỗi nghiệp vụ"
                    },
                    "403": {
                        "description": "Địa chỉ không thuộc về user này"
                    },
                    "409": {
                        "description": "Tồn kho không đủ"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Nhập Token theo định dạng: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}